requirements.txt
selenium==3.141.0
mysql_connector_repackaged==0.3.1
